refactor: consolidate PhysicalTenantContext into the authentication module
The physical-tenant request-context holder is used both by the gateway routing layer
(gateway-rest, gateway-mcp) and by the authentication module (PhysicalTenantPreSecurityFilter,
BasicAuthUserDetailsAdapter). It lived in gateway-mapping-http, but the authentication module cannot
depend on gateway-mapping-http (the dependency runs the other way), so the authentication code
carried a byte-identical copy. Because the request-attribute key derives from the class FQN, the two
copies used different keys — a latent cross-wiring hazard.

Make io.camunda.authentication.pt.PhysicalTenantContext the single canonical class (authentication
is the lowest module every gateway module already depends on), delete the gateway-mapping-http copy,
move its test, and repoint all gateway-rest / gateway-mcp / dist imports to it.

gateway-mcp now declares camunda-authentication directly (transitive deps excluded, matching the
existing pattern) since it uses the class directly; gateway-rest and dist already did. spring-test is
dropped from gateway-mapping-http as its only consumer (the context test) moved out.
