--- gateways/gateway-mapping-http/src/main/java/io/camunda/gateway/mapping/http/physicaltenants/PhysicalTenantContext.java
@@ -1,75 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.gateway.mapping.http.physicaltenants;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.web.context.request.RequestAttributes;
-import org.springframework.web.context.request.RequestContextHolder;
-
-/**
- * Carries the resolved physical tenant id for the current request.
- *
- * <p>Used by both the REST gateway ({@code /physical-tenants/{physicalTenantId}/v2/...} routes) and
- * the MCP gateway ({@code /physical-tenants/{physicalTenantId}/mcp/...} routes). The id is
- * populated on the request by the respective interceptor and read downstream via {@link #current()}
- * (request-bound thread) or {@link #getPhysicalTenantId(HttpServletRequest)}.
- */
-public final class PhysicalTenantContext {
-
-  /** Default physical tenant id used when no prefix is present in the request path. */
-  public static final String DEFAULT_PHYSICAL_TENANT_ID = "default";
-
-  /** URL path prefix for physical-tenant-scoped routes (e.g. {@code /physical-tenants/foo/}). */
-  public static final String PHYSICAL_TENANTS_PATH_SEGMENT = "/physical-tenants/";
-
-  /** URI template variable carrying the physical tenant id in prefixed routes. */
-  public static final String PATH_VARIABLE_PHYSICAL_TENANT_ID = "physicalTenantId";
-
-  /**
-   * URI template prefix for physical-tenant-scoped routes (e.g. {@code
-   * /physical-tenants/{physicalTenantId}/...}).
-   */
-  public static final String PHYSICAL_TENANT_URI_PREFIX =
-      PHYSICAL_TENANTS_PATH_SEGMENT + "{" + PATH_VARIABLE_PHYSICAL_TENANT_ID + "}";
-
-  /** Request attribute key under which the resolved id is stored. */
-  public static final String REQUEST_ATTRIBUTE_PHYSICAL_TENANT_ID =
-      PhysicalTenantContext.class.getName() + ".PHYSICAL_TENANT_ID";
-
-  private PhysicalTenantContext() {}
-
-  public static void setPhysicalTenantId(
-      final HttpServletRequest request, final String physicalTenantId) {
-    request.setAttribute(REQUEST_ATTRIBUTE_PHYSICAL_TENANT_ID, physicalTenantId);
-  }
-
-  public static String getPhysicalTenantId(final HttpServletRequest request) {
-    return asString(request.getAttribute(REQUEST_ATTRIBUTE_PHYSICAL_TENANT_ID));
-  }
-
-  /**
-   * @return the resolved physical tenant id for the request being processed on this thread, or
-   *     {@link #DEFAULT_PHYSICAL_TENANT_ID} when no request context is bound or the attribute is
-   *     not set.
-   */
-  public static String current() {
-    final RequestAttributes attributes = RequestContextHolder.getRequestAttributes();
-    if (attributes == null) {
-      return DEFAULT_PHYSICAL_TENANT_ID;
-    }
-    final String value =
-        asString(
-            attributes.getAttribute(
-                REQUEST_ATTRIBUTE_PHYSICAL_TENANT_ID, RequestAttributes.SCOPE_REQUEST));
-    return value != null ? value : DEFAULT_PHYSICAL_TENANT_ID;
-  }
-
-  private static String asString(final Object value) {
-    return value != null ? value.toString() : null;
-  }
-}
