+++ gateways/gateway-mcp/src/main/java/io/camunda/gateway/mcp/config/CamundaMcpServersAutoConfiguration.java
@@ -7,11 +7,11 @@
  */
 package io.camunda.gateway.mcp.config;
 
-import static io.camunda.gateway.mapping.http.physicaltenants.PhysicalTenantContext.PATH_VARIABLE_PHYSICAL_TENANT_ID;
-import static io.camunda.gateway.mapping.http.physicaltenants.PhysicalTenantContext.PHYSICAL_TENANT_URI_PREFIX;
+import static io.camunda.authentication.pt.PhysicalTenantContext.PATH_VARIABLE_PHYSICAL_TENANT_ID;
+import static io.camunda.authentication.pt.PhysicalTenantContext.PHYSICAL_TENANT_URI_PREFIX;
 
 import io.camunda.configuration.api.physicaltenants.PhysicalTenantIds;
-import io.camunda.gateway.mapping.http.physicaltenants.PhysicalTenantContext;
+import io.camunda.authentication.pt.PhysicalTenantContext;
 import io.camunda.gateway.mcp.ConditionalOnMcpGatewayEnabled;
 import io.camunda.gateway.mcp.config.server.RequestHandlerCustomizer;
 import io.camunda.gateway.mcp.config.server.ToolRepository;
