+++ gateways/gateway-mcp/src/test/java/io/camunda/gateway/mcp/config/CamundaMcpServersAutoConfigurationTenantFilterTest.java
@@ -14,7 +14,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import io.camunda.configuration.api.physicaltenants.PhysicalTenantIds;
-import io.camunda.gateway.mapping.http.physicaltenants.PhysicalTenantContext;
+import io.camunda.authentication.pt.PhysicalTenantContext;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.ObjectProvider;
