+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/mapper/PhysicalTenantRequestMappingHandlerMappingTest.java
@@ -9,7 +9,7 @@ package io.camunda.zeebe.gateway.rest.mapper;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.camunda.gateway.mapping.http.physicaltenants.PhysicalTenantContext;
+import io.camunda.authentication.pt.PhysicalTenantContext;
 import io.camunda.zeebe.gateway.rest.annotation.ClusterScoped;
 import io.camunda.zeebe.gateway.rest.controller.CamundaRestController;
 import java.lang.reflect.Method;
